Add baremetal _nostd variant
Note that product_available and vendor_available are dropped from the
existing target because they are irrelevant for a _host target (as per
the change in aosp/2763825).

Add a visibility restriction to the new variant to preserve the
limitation that this crate should not be used in general Android code.

Bug: 303192701
Test: build both variants
Change-Id: I36d5dc3d6a22a750963b0ffb592f65880ae60ec0